requirements.txt
streamlit
pandas
sqlalchemy
pymongo
python-dotenv